+++ src/components/flux/config/CountSelector.vue
@@ -6,7 +6,7 @@
         <info-icon :content="$t('flux.description.numbers')" />
       </div>
       <div class="flex justify-end items-center">
-        <el-input-number v-model="value" size="small" controls-position="right" />
+        <el-input-number v-model="value" controls-position="right" />
       </div>
     </div>
     <div class="w-full">
@@ -15,14 +15,14 @@
   </div>
 </template>
 
-<script>
+<script lang="ts">
 import { defineComponent } from 'vue';
 import { ElSlider, ElInputNumber } from 'element-plus';
 import InfoIcon from '@/components/common/InfoIcon.vue';
-import { FLUX_DEFAULT_NUMBERS } from '@/constants';
+import { FLUX_DEFAULT_COUNT } from '@/constants';
 
 export default defineComponent({
-  name: 'NumbersSelector',
+  name: 'CountSelector',
   components: {
     ElSlider,
     InfoIcon,
@@ -31,20 +31,20 @@ export default defineComponent({
   computed: {
     value: {
       get() {
-        return this.$store.state.flux?.config?.numbers;
+        return this.$store.state.flux?.config?.count;
       },
-      set(val) {
-        console.debug('set stylize', val);
+      set(val: number) {
+        console.debug('set count', val);
         this.$store.commit('flux/setConfig', {
           ...this.$store.state.flux?.config,
-          numbers: val
+          count: val
         });
       }
     }
   },
   mounted() {
     if (!this.value) {
-      this.value = FLUX_DEFAULT_NUMBERS;
+      this.value = FLUX_DEFAULT_COUNT;
     }
   }
 });
