+++ src/constants/flux.ts
@@ -1,6 +1,6 @@
 export const FLUX_SERVICE_ID = 'e8ebe055-429a-48e9-9d34-5089eae5fd89';
 
-export const FLUX_DEFAULT_MODEL = 'flux';
+export const FLUX_DEFAULT_MODEL = 'flux-dev';
 export const FLUX_DEFAULT_ASPECT_RATIO = '1:1';
-export const FLUX_DEFAULT_NUMBERS = 1;
+export const FLUX_DEFAULT_COUNT = 1;
 export const FLUX_DEFAULT_QUALITY = 80;
